#2424 rename and move to the eo-maven-package
