+++ .codacy.yml
@@ -27,5 +27,5 @@
 exclude_paths:
   - "src/test/groovy/*.groovy"
   - "eo-runtime/src/test/java/integration/PhiUnphiIT.java"
-  - "eo-maven-plugin/src/test/java/integration/MojasIT.java"
+  - "eo-maven-plugin/src/test/java/integration/AssembleMojoIT.java"
   - "eo-maven-plugin/src/test/java/org/eolang/maven/LintMojoTest.java"
