+++ eo-maven-plugin/src/test/java/org/eolang/maven/AssembleMojoIT.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package integration;
+package org.eolang.maven;
 
 import com.yegor256.Mktmp;
 import com.yegor256.MktmpResolver;
@@ -30,12 +30,7 @@ import com.yegor256.farea.Farea;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
-import org.eolang.maven.AssembleMojo;
-import org.eolang.maven.ContainsFiles;
-import org.eolang.maven.ParseMojo;
-import org.eolang.maven.PullMojo;
-import org.eolang.maven.ResolveMojo;
-import org.eolang.maven.ShakeMojo;
+
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
@@ -49,7 +44,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 @SuppressWarnings({"JTCOP.RuleAllTestsHaveProductionClass", "JTCOP.RuleNotContainsTestWord"})
 @ExtendWith(WeAreOnline.class)
 @ExtendWith(MktmpResolver.class)
-final class MojasIT {
+final class AssembleMojoIT {
 
     @Test
     void assemblesTogether(@Mktmp final Path temp) throws IOException {
@@ -64,7 +59,7 @@ final class MojasIT {
                     .append("org.eolang", "eo-runtime", "0.51.6");
                 f.files()
                     .file("src/main/eo/one/main.eo")
-                    .write(MojasIT.helloWorld().getBytes(StandardCharsets.UTF_8));
+                    .write(AssembleMojoIT.helloWorld().getBytes(StandardCharsets.UTF_8));
                 f.build()
                     .plugins()
                     .appendItself()
@@ -158,7 +153,7 @@ final class MojasIT {
                 f.clean();
                 f.files()
                     .file("src/main/eo/one/main.eo")
-                    .write(MojasIT.helloWorld().getBytes(StandardCharsets.UTF_8));
+                    .write(AssembleMojoIT.helloWorld().getBytes(StandardCharsets.UTF_8));
                 f.build()
                     .plugins()
                     .appendItself()
